Extract shared parser methods into a RestParser
