+++ lib/wpclient/media_parser.rb
@@ -1,5 +1,7 @@
 module Wpclient
   class MediaParser
+    include RestParser
+
     def self.parse(data)
       new(data).to_media
     end
@@ -35,17 +37,8 @@ module Wpclient
     end
 
     def assign_rendered(media)
-      media.title = data.fetch("title", {}).fetch("rendered")
-      media.guid = data.fetch("guid", {}).fetch("rendered")
-    end
-
-    def read_date(name)
-      # Try to read UTC time first
-      if (gmt_time = data["#{name}_gmt"])
-        Time.iso8601("#{gmt_time}Z")
-      elsif (local_time = data[name])
-        Time.iso8601(local_time)
-      end
+      media.title = rendered("title")
+      media.guid = rendered("guid")
     end
   end
 end
